pyproject.toml
[project]
name = "ge-test"
version = "0.1.0"
description = "GE Test"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.11.18",
    "fastapi[standard]>=0.115.12",
    "lxml-html-clean>=0.4.2",
    "openai>=1.81.0",
    "passlib>=1.7.4",
    "pydantic>=2.11.4",
    "pydantic-settings>=2.9.1",
    "pyjwt>=2.10.1",
    "pytest>=8.3.5",
    "pytest-asyncio>=0.26.0",
    "requests-html>=0.10.0",
    "sqlmodel>=0.0.24",
]
